requirements.txt
Pillow==9.5.0
psd_tools==1.9.27
